flytt funksjoner for å oppdatere og slette endret utbetaling inn i context
